Prevent user from paying cart twice
Implements a guard clause to prevent paying a cart more than once

+++ src/main/kotlin/com/jumarket/selfcheckout/services/impl/CartService.kt
@@ -3,15 +3,20 @@ package com.jumarket.selfcheckout.services.impl
 import com.jumarket.selfcheckout.entities.Cart
 import com.jumarket.selfcheckout.entities.CartItem
 import com.jumarket.selfcheckout.enumerations.PaymentMethod
+import com.jumarket.selfcheckout.exceptions.CartAlreadyPaidException
+import com.jumarket.selfcheckout.exceptions.CartNotFoundException
+import com.jumarket.selfcheckout.exceptions.ProductAlreadyOnCartException
 import com.jumarket.selfcheckout.repositories.CartItemRepository
 import com.jumarket.selfcheckout.repositories.CartRepository
 import com.jumarket.selfcheckout.services.ICartService
+import org.springframework.data.domain.Example
+import org.springframework.data.domain.ExampleMatcher
 import org.springframework.stereotype.Service
 
 @Service
 class CartService(
         private val cartRepository: CartRepository,
-        private val cartItemRepository: CartItemRepository
+        private val cartItemRepository: CartItemRepository,
 ) : ICartService {
 
     override fun createCart(): Cart {
@@ -20,23 +25,38 @@ class CartService(
 
     override fun findCart(id: Long): Cart {
         return cartRepository.findById(id).orElseThrow {
-            throw IllegalArgumentException("Cart with id $id not found")
+            throw CartNotFoundException("Cart with id $id not found")
         }
     }
 
     override fun addCartItem(cartItem: CartItem) {
+        if (cartItem.cart!!.paid) throw CartAlreadyPaidException("Can not add itens to a paid cart")
+        if (isProductOnCart(cartItem))
+                throw ProductAlreadyOnCartException(
+                        "Product ${cartItem.product?.productName} already on cart"
+                )
         cartItemRepository.save(cartItem)
     }
 
     override fun removeCartItem(cartItem: CartItem) {
+        if (cartItem.cart!!.paid)
+                throw CartAlreadyPaidException("Can not remove itens of a paid cart")
         cartItemRepository.delete(cartItem)
     }
 
     override fun payCart(cart: Cart, method: PaymentMethod) {
+        if (cart.paid) throw CartAlreadyPaidException("Can not pay a cart twice")
         cart.apply {
             paid = true
             paymentMethod = method
         }
         cartRepository.save(cart)
+    }
+
+    fun isProductOnCart(cartItem: CartItem): Boolean {
+        val matcher = ExampleMatcher.matching().withIgnorePaths("quantity", "totalPrice")
+        val cartItemExample =
+                Example.of(CartItem(cart = cartItem.cart, product = cartItem.product), matcher)
+        return cartItemRepository.findOne(cartItemExample).isPresent()
     }
 }
